fix: harden integer widths at five allocation boundaries
A sweep of the remaining c_int/u32 sizing sites flagged by the audit. All
six were verified before being changed; the corpus is byte-identical, so
the sweep is output-neutral.

md4x-heal.zig buf_append is the sharp one. `buf.size +% len > buf.cap`
used wrapping arithmetic to guard against a wrap, so past 2^32 the test
read false, no realloc happened, and the @memcpy wrote past a buffer that
was never grown. `+%` is well defined even in ReleaseSafe, so no build
mode caught it. Now a checked add latches buf.err, which md_heal already
turns into a clean -1. buf_init had the same wrap for an input within 64
bytes of 4 GiB.

md_analyze_table_alignment's dash scan is now bounded by the row end
(md4c ecbb091). It is unreachable today -- and commit 1094346 strengthens
that, since TABLE_MAXCOLCOUNT is now exactly maxInt(u16) and the count is
stored in a u16, so n_align cannot exceed the groups present -- but
ctx.ch() indexes a [*c] pointer that Zig bounds-checks in no build mode,
so the failure would be a silent over-read rather than a panic.

The ref-def hashtable size is computed as n + n/4 rather than n*5/4, which
is the same value for every n but never forms the 5x intermediate, so it
cannot wrap. The complex-bucket growth one level down is refused above
maxInt(c_int). MD_REF_DEF_LIST.alloc_ref_defs keeps its type: the extern
struct's layout is frozen by its flexible array.

MD_ATTRIBUTE_BUILD's counters widen to usize and refuse at the opener
above a cap justified by a substring costing at least one byte of raw
attribute text, which is itself an SZ.

Oversized input is refused rather than truncated: the CLI above
maxInt(MD_SIZE), and NAPI at all five call sites that duplicated the
pattern, before the malloc rather than after.

The containers.items.len decrement is guarded. No repro was found and the
block layer is believed to balance openers and closers, but the counter is
a usize, so a lone closer would wrap it and the subsequent index would
read out of a nonsense slice in ReleaseFast.

Boundaries too large to reach with real input are pinned by unit tests
that seed the counters at the exact overflow value and observe the
requested size through a refusing allocator.

+++ src/parser/types.zig
@@ -353,7 +353,10 @@ pub const MD_CTX = struct {
     // built only after collection completes — no append moves the buffer after.)
     ref_defs: std.ArrayListUnmanaged(MD_REF_DEF) = .empty,
     ref_def_hashtable: [*c]?*anyopaque = null,
-    ref_def_hashtable_size: c_int = 0,
+    // A bucket count, so `usize` — the type the allocation and every index into
+    // the table already want. md4c keeps it `int` and computes `n * 5 / 4` in
+    // `int`, which overflows above ~429M ref-defs; upstream 19dd06f widened it.
+    ref_def_hashtable_size: usize = 0,
     max_ref_def_output: SZ = 0,
 
     // Stack of inline/span markers. (PLAN 8.1: ArrayListUnmanaged. The emphasis
